Align audit action filters
Keep the audit action allowlist in one shared schema so the UI
and API cannot drift apart. This prevents valid session-related
filters from failing in the audit log search endpoint.

Refs #113
